requirements.txt
openai>=1.52.0
gspread>=6.1.2
google-auth>=2.34.0
tenacity>=9.0.0
python-dotenv>=1.0.1
